pyproject.toml
[tool.poetry]
name = "prefect_processor"
version = "0.0.1"
description = "Transcript the audio file based on the channel"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
requests = "^2.32"
prefect = {extras = ["kubernetes"], version = "^3.4.4"}
ffmpeg-python = "^0.2.0"
pydub = "^0.25.1"
openai-whisper = "^20240930"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
pytest = "^8.0"
